actually fixed the game image problem
first thing i did was make sure the cachedImages directory was actually being created. this involved modifying the code in Add&EditGameView. then I figured out that in LogUtils there was code to delete the oldest logs if there are more than 3. when 3 logs were created the cachedImages folder was deleted. now it checks to make sure whatever if being deleted has the .log extension.

Signed-off-by: James Hughes <github@jxhug.me>